collector-opi-deprecated/requirements.txt
python-dotenv==0.21.0
pathlib==1.0.1
pyA20==0.2.1
pytest==8.4.2
